+++ src/demo_factories/demo_factories.py
@@ -1,0 +1,4 @@
+import dipdup
+
+# NOTE: Allows to use the package as a working directory
+dipdup.workdir(__file__, __name__)